pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcybox-sqlbuddy"
version = "0.1.0"
description = "AI-powered SQL query generator using OpenAI GPT or Anthropic Claude. Automatic schema extraction, query validation, and optimization for MySQL and PostgreSQL."
readme = "README.md"
authors = [{ name = "Mister_iks", email = "shadow@pcybox.com" }]
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = [
    "sql",
    "ai",
    "llm",
    "openai",
    "claude",
    "gpt",
    "mysql",
    "postgresql",
    "query-generator",
    "natural-language",
    "database",
    "sql-generator",
    "text-to-sql",
]

dependencies = [
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.9",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/Mister-iks/pcybox-sqlbuddy"
Documentation = "https://github.com/Mister-iks/pcybox-sqlbuddy#readme"
Repository = "https://github.com/Mister-iks/pcybox-sqlbuddy"
Issues = "https://github.com/Mister-iks/pcybox-sqlbuddy/issues"

[project.scripts]
sqlbuddy = "sqlbuddy.cli:main"
